Fixed potential errors

+++ src/thrift/protocol/compact_stream_write.rs
@@ -1,20 +1,3 @@
-// Licensed to the Apache Software Foundation (ASF) under one
-// or more contributor license agreements. See the NOTICE file
-// distributed with this work for additional information
-// regarding copyright ownership. The ASF licenses this file
-// to you under the Apache License, Version 2.0 (the
-// "License"); you may not use this file except in compliance
-// with the License. You may obtain a copy of the License at
-//
-//   http://www.apache.org/licenses/LICENSE-2.0
-//
-// Unless required by applicable law or agreed to in writing,
-// software distributed under the License is distributed on an
-// "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY
-// KIND, either express or implied. See the License for the
-// specific language governing permissions and limitations
-// under the License.
-
 use std::convert::From;
 
 use async_trait::async_trait;
@@ -81,7 +64,7 @@ where
     async fn write_list_set_begin(
         &mut self,
         element_type: TType,
-        element_count: i32,
+        element_count: u32,
     ) -> Result<usize> {
         let elem_identifier = collection_type_to_u8(element_type);
         if element_count <= 14 {
@@ -92,12 +75,7 @@ where
 
             let header = 0xF0 | elem_identifier;
             written += self.write_byte(header).await?;
-            // element count is strictly positive as per the spec, so
-            // cast i32 as u32 so that varint writing won't use zigzag encoding
-            written += self
-                .transport
-                .write_varint_async(element_count as u32)
-                .await?;
+            written += self.transport.write_varint_async(element_count).await?;
             Ok(written)
         }
     }
